Refactor: Split AuthContext into smaller files
This commit refactors the AuthContext.tsx file into smaller, more manageable files to improve code organization and readability.
